Update soap spec to test actual behaviour
There were 5 examples here that were all testing the same method being
called with different attributes that were expected to be transformed.
The code that would do this transformation is commented out, so these
tests are redundant.

This leaves a single example that tests the properties we pass to
`soap_post` get included as-is in the request without transformation.